implement execute and memory for each instruction handle memory read/write

+++ src/Instructions/Operands/ImmediateOperand.java
@@ -1,10 +1,10 @@
 package Instructions.Operands;
 
 public class ImmediateOperand extends Operand{
-	public int value;
+	public double value;
 	
 	public ImmediateOperand(String raw_value) {
 		super();
-		this.value = Integer.parseInt(raw_value);
+		this.value = Float.parseFloat(raw_value);
 	}
 }
